WIP: start-up modifications
