+++ src/LodeRunner.API.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 16.0.31515.178
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LodeRunner.API", "LodeRunner.API\LodeRunner.API.csproj", "{6D327B55-A1EB-49DD-A9E3-182838D5841C}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LodeRunner.Data", "LodeRunner.Data\LodeRunner.Data.csproj", "{45400898-CFE7-4198-BAC1-471F59E2A106}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -15,6 +17,10 @@ Global
 		{6D327B55-A1EB-49DD-A9E3-182838D5841C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6D327B55-A1EB-49DD-A9E3-182838D5841C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6D327B55-A1EB-49DD-A9E3-182838D5841C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{45400898-CFE7-4198-BAC1-471F59E2A106}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{45400898-CFE7-4198-BAC1-471F59E2A106}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{45400898-CFE7-4198-BAC1-471F59E2A106}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{45400898-CFE7-4198-BAC1-471F59E2A106}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
